requirements.txt
#  pip install -r requirements.txt
pandas-1.3.1
Django-3.2.5